# swagger.yaml
openapi: 3.0.0
info:
  title: Contoh Microservice API
  version: 1.0.0
  description: Ini adalah contoh dokumentasi untuk API microservice.
  contact:
    email: dimaspondrao@gmail.com
    name: Dimas Pondra Oktafianto
    url: https://dimasoktafianto.my.id
servers:
  - url: https://microservice.dimasoktafianto.my.id
    variables: {}
    description: Server Production
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    401Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            type: object
            properties:
              status:
                type: string
                example: error
              message:
                type: string
                example: Unauthorized, token is invalid.
    403Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            type: object
            properties:
              status:
                type: string
                example: error
              message:
                type: string
                example: Forbidden, access denied.
...
    422MyCourse:
      description: Unprocessable Entity
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: The course id field is required.
              errors:
                type: object
                properties:
                  course_id:
                    type: array
                    items:
                      example: The course id field is required.
...
  requestBodies:
...
    requestBodyMyCourse:
      type: object
      properties:
        course_id:
          type: integer
          example: 1
      required:
        - course_id
paths:
...
  /my-courses:
    summary: My course
    get:
      summary: List my courses
      description: Get all data of my courses
      operationId: getMyCourses
      parameters:
        - name: include
          in: query
          description: Adding data relations
          required: true
          schema:
            type: string
            default: course
            example: course
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  message:
                    type: string
                    example: Successfully load data.
                  data:
                    type: array
                    items:
                      oneOf:
                        - type: object
                          properties:
                            id:
                              type: integer
                              example: 1
                            user:
                              type: object
                              properties:
                                id:
                                  type: integer
                                  example: 3
                                name:
                                  type: string
                                  example: Dimas
      tags:
        - My Course
    post:
      summary: Create my course
      description: Create new my course
      operationId: postMyCourse
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/requestBodies/requestBodyMyCourse'
      security:
        - bearerAuth: []
      responses:
        '401':
          $ref: '#/components/responses/401Unauthorized'
        '403':
          $ref: '#/components/responses/403Forbidden'
        '422':
          $ref: '#/components/responses/422MyCourse'
      tags:
        - My Course
...